# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Deetoo API
  version: 1.0.0
  description: 'Deetoo multi-vendor food delivery platform API (Kenya/East Africa).


    **Coverage note (honest, not aspirational):** this file was generated from the actual Express router
    source (scripts/openapi-coverage-check.ts enforces this in CI going forward -- it fails the build
    if a route exists in code but not here, or vice versa), so every path, method, tag, and auth requirement
    below is real. Of the 261 operations, 31 core commerce-flow operations (public discovery, auth, cart,
    checkout, order, payment) have hand-verified request/response schemas transcribed from the actual
    Zod validators and domain types. The remaining operations are marked `STUB` in their summary: they
    are registered with the correct path/method/tag/security so drift is still caught, but their request/response
    bodies are intentionally left as a generic envelope rather than guessed. Replacing each STUB with
    its real schema is tracked as ongoing work, not represented here as already done.'
  contact:
    name: Deetoo Engineering
servers:
- url: http://localhost:3000/api/v1
  description: Local development
...
security:
- bearerAuth: []
paths:
...
  /admin/orders:
    get:
      summary: STUB - GET /admin/orders (registered, not yet individually specified)
      tags:
      - Admin
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: admin_get_admin_orders
      security:
      - bearerAuth: []
...
  /cart:
    get:
      summary: Get the current customer's cart
      tags:
      - Cart
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: cart_get_cart
      security:
      - bearerAuth: []
    delete:
      summary: Clear the cart
      tags:
      - Cart
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: cart_delete_cart
      security:
      - bearerAuth: []
...
  /customer/orders:
    get:
      summary: STUB - GET /customer/orders (registered, not yet individually specified)
      tags:
      - Customer
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: customer_get_customer_orders
      security:
      - bearerAuth: []
...
  /merchant:
    get:
      summary: STUB - GET /merchant (registered, not yet individually specified)
      tags:
      - Merchant Orders
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: merchant-orders_get_merchant
      security:
      - bearerAuth: []
...
  /orders:
    post:
      summary: Create an order from a bound checkout quote (idempotent via Idempotency-Key header)
      tags:
      - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: order_post_orders
      security:
      - bearerAuth: []
...
  /finance/disbursements:
    get:
      summary: List durable money-out attempts
      tags: [Finance]
      responses:
        '200':
          description: Disbursement attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_disbursements
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session access token issued by POST /auth/login or /auth/register/customer.
    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: Shared secret injected by the trusted ingress gateway after it independently validates
        the call genuinely originated from the provider (M-PESA Daraja). This is gateway authentication,
        not a provider-issued signature -- Safaricom does not sign these requests itself. Stripe webhooks
        instead use Stripe-Signature HMAC verification against the raw body and are not covered by this
        scheme.
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing, invalid, or expired credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiErrorDetail:
      type: object
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          description: Stable machine-readable error code, e.g. ORDER_NOT_FOUND
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
    ApiError:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDetail'
    PaginationMeta:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        total_count:
          type: integer
    ApiResponseGeneric:
      type: object
      required:
      - data
      description: Generic success envelope. data is intentionally untyped here where a route has not
        yet had its specific response schema individually verified (see the STUB markers in the paths
        above).
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
        success:
          type: boolean
    MoneyMinor:
      type: integer
      description: 'Money convention used throughout this API: every amount is an integer count of the
        currency''s minor unit (e.g. 125000 = KES 1,250.00), always paired with a sibling `currency` field
        (ISO 4217, e.g. "KES"). There is no floating-point money anywhere in this API. Fields follow the
        `<name>_minor` naming pattern (e.g. subtotal_minor, total_minor, amount_minor).'
...
    Order:
      type: object
      required:
      - id
      - public_code
      - order_number
      - customer_id
      - branch_id
      - merchant_id
      - status
      - currency
      - subtotal_minor
      - delivery_fee_minor
      - service_fee_minor
      - discount_minor
      - total_minor
      - items
      properties:
        id:
          type: string
        public_code:
          type: string
        order_number:
          type: string
        checkout_quote_id:
          type: string
        customer_id:
          type: string
        branch_id:
          type: string
        merchant_id:
          type: string
        status:
          type: string
          description: See DEE-STATE-001 for the full order state machine.
        currency:
          type: string
        subtotal_minor:
          $ref: '#/components/schemas/MoneyMinor'
        delivery_fee_minor:
          $ref: '#/components/schemas/MoneyMinor'
        service_fee_minor:
          $ref: '#/components/schemas/MoneyMinor'
        discount_minor:
          $ref: '#/components/schemas/MoneyMinor'
        total_minor:
          $ref: '#/components/schemas/MoneyMinor'
        items:
          type: array
          items:
            type: object
        special_instructions:
          type: string
          nullable: true
        placed_at:
          type: string
          format: date-time
          nullable: true
...
    CartResponse:
      allOf:
      - $ref: '#/components/schemas/ApiResponseGeneric'
      properties:
        data:
          type: object
          description: Cart
          properties:
            items:
              type: array
              items:
                type: object
            subtotal_minor:
              $ref: '#/components/schemas/MoneyMinor'
            branch_id:
              type: string
...
    OrderResponse:
      allOf:
      - $ref: '#/components/schemas/ApiResponseGeneric'
      properties:
        data:
          $ref: '#/components/schemas/Order'
...
    CreateOrderRequest:
      type: object
      description: One of quote_id / quoteId is required (accepts either casing).
      properties:
        quote_id:
          type: string
        quoteId:
          type: string
        special_instructions:
          type: string
          maxLength: 500
...